Passing Unit Tests

+++ src/DelMesh/Point2D.cpp
@@ -18,7 +18,7 @@ std::ostream& operator <<(std::ostream& os, const Point2D& pt){
 
 
 //Euclidian distance from another point
-float Point2D::dist(Point2D other){
+double Point2D::dist(Point2D other){
     return sqrt( pow(V(0) - other.V(0),2) + pow(V(1) - other.V(1),2));
 }
